Add log in functionality
##Summary

In order to allow our users to see the app and use our
database we add the authentication with google. Maybe
waaaaay later we will add other authentication.
This was added so that the database does not have to be
in test mode which is insecure but rather in normal dev
mode.

## Test Plan
* Open the app
* click log in
* select my google account
* click log in
 :-O
it worked!
* click log out
 :-O
it worked, again!